+++ src/index.js
@@ -1,0 +1,10 @@
+const { registerOpenBox } = require('./commands/openBox')
+
+function activate(context) {
+	const disposable = registerOpenBox(context)
+	context.subscriptions.push(disposable)
+}
+
+function deactivate() {}
+
+module.exports = { activate, deactivate }